fix-issues-in-homework8
